% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisp.R
\name{lisp}
\alias{lisp}
\title{local indicator of stratified power}
\usage{
lisp(
  formula,
  data,
  bandwidth = NULL,
  discvar = NULL,
  discnum = 3:8,
  discmethod = c("sd", "equal", "geometric", "quantile", "natural"),
  cores = 1,
  ...
)
}
\arguments{
\item{formula}{A formula.}

\item{data}{An \code{sf} object of observation data.}

\item{bandwidth}{(optional) The bandwidth employed to select "local" data.}

\item{discvar}{(optional) Name of continuous variable columns that need to be discretized. Noted
that when \code{formula} has \code{discvar}, \code{data} must have these columns. By default, all independent
variables are used as \code{discvar}.}

\item{discnum}{(optional) A vector of number of classes for discretization. Default is \code{3:8}.}

\item{discmethod}{(optional) A vector of methods for discretization, default is using
\code{c("sd","equal","geometric","quantile","natural")} by invoking \code{sdsfun}.}

\item{cores}{(optional) Positive integer (default is 1). When cores are greater than 1, use
multi-core parallel computing.}

\item{...}{(optional) Other arguments passed to \code{gdverse::gd_opttunidisc()}. A useful parameter
is \code{seed}, which is used to set the random number seed.}
}
\value{
A \code{tibble}.
}
\description{
local indicator of stratified power
}
\examples{
gtc = readr::read_csv(system.file("extdata/gtc.csv", package = "localsp"))
gtc
gtc = sf::st_as_sf(gtc, coords = c("X","Y"), crs = 4326)
gtc

\dontrun{
## The following code takes approximately 5 minutes to run:
lisp(GTC ~ ., data = gtc, bandwidth = 618295.4, cores = 6)
}
}
